Update if arguments

+++ nextcloud-startup-script.sh
@@ -42,14 +42,10 @@ network_ok() {
     fi
 }
 
-
 # DEBUG mode
 if [ $DEBUG -eq 1 ]
 then
-    set -e
-    set -x
-else
-    sleep 1
+    set -ex
 fi
 
 # Check if root
@@ -124,10 +120,8 @@ then
 else
     wget -q $STATIC/passman.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/passman.sh ]
+if [ ! -f $SCRIPTS/passman.sh ]
 then
-    sleep 0.1
-else
     echo "passman failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -141,10 +135,8 @@ then
 else
     wget -q $STATIC/nextant.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/nextant.sh ]
+if [ ! -f $SCRIPTS/nextant.sh ]
 then
-    sleep 0.1
-else
     echo "nextant failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again." 
     exit 1
@@ -158,10 +150,8 @@ then
 else
     wget -q $STATIC/collabora.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/collabora.sh ]
+if [ ! -f $SCRIPTS/collabora.sh ]
 then
-    sleep 0.1
-else
     echo "collabora failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -175,10 +165,8 @@ then
 else
     wget -q $STATIC/spreedme.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/spreedme.sh ]
+if [ ! -f $SCRIPTS/spreedme.sh ]
 then
-    sleep 0.1
-else
     echo "spreedme failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -192,10 +180,8 @@ then
 else
     wget -q $STATIC/temporary-fix.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/temporary-fix.sh ]
+if [ ! -f $SCRIPTS/temporary-fix.sh ]
 then
-    sleep 0.1
-else
     echo "temporary-fix failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -209,10 +195,8 @@ then
 else
     wget -q $STATIC/security.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/security.sh ]
+if [ ! -f $SCRIPTS/security.sh ]
 then
-    sleep 0.1
-else
     echo "security failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -226,10 +210,8 @@ then
 else
     wget -q $STATIC/update.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/update.sh ]
+if [ ! -f $SCRIPTS/update.sh ]
 then
-    sleep 0.1
-else
     echo "nextcloud_update failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -243,10 +225,8 @@ then
 else
     wget -q $STATIC/phpmyadmin_install_ubuntu16.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/phpmyadmin_install_ubuntu16.sh ]
+if [ ! -f $SCRIPTS/phpmyadmin_install_ubuntu16.sh ]
 then
-    sleep 0.1
-else
     echo "phpmyadmin_install_ubuntu16 failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -260,10 +240,8 @@ then
 else
     wget -q $STATIC/update-config.php -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/update-config.php ]
+if [ ! -f $SCRIPTS/update-config.php ]
 then
-    sleep 0.1
-else
     echo "update-config failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -277,10 +255,8 @@ then
 else
     wget -q $LETS_ENC/activate-ssl.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/activate-ssl.sh ]
+if [ ! -f $SCRIPTS/activate-ssl.sh ]
 then
-    sleep 0.1
-else
     echo "activate-ssl failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -294,10 +270,8 @@ then
 else
     wget -q $STATIC/trusted.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/trusted.sh ]
+if [ ! -f $SCRIPTS/trusted.sh ]
 then
-    sleep 0.1
-else
     echo "trusted failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -311,10 +285,8 @@ then
 else
     wget -q $STATIC/ip.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/ip.sh ]
+if [ ! -f $SCRIPTS/ip.sh ]
 then
-    sleep 0.1
-else
     echo "ip failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -328,10 +300,8 @@ then
 else
     wget -q $STATIC/test_connection.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/test_connection.sh ]
+if [ ! -f $SCRIPTS/test_connection.sh ]
 then
-    sleep 0.1
-else
     echo "test_connection failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -345,10 +315,8 @@ then
 else
     wget -q $STATIC/setup_secure_permissions_nextcloud.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/setup_secure_permissions_nextcloud.sh ]
+if [ ! -f $SCRIPTS/setup_secure_permissions_nextcloud.sh ]
 then
-    sleep 0.1
-else
     echo "setup_secure_permissions_nextcloud failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -362,10 +330,8 @@ then
 else
     wget -q $STATIC/change_mysql_pass.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/change_mysql_pass.sh ]
+if [ ! -f $SCRIPTS/change_mysql_pass.sh ]
 then
-    sleep 0.1
-else
     echo "change_mysql_pass failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -379,10 +345,8 @@ then
 else
     wget -q $STATIC/nextcloud.sh -P $SCRIPTS
 fi
-if [ -f $SCRIPTS/nextcloud.sh ]
+if [ ! -f $SCRIPTS/nextcloud.sh ]
 then
-    sleep 0.1
-else
     echo "nextcloud failed"
     echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
     exit 1
@@ -394,6 +358,12 @@ then
     wget -q $GITHUB_REPO/index.php -P $SCRIPTS
 else
     wget -q $GITHUB_REPO/index.php -P $SCRIPTS
+fi
+if [ ! -f $SCRIPTS/index.php ]
+then
+    echo "index.php failed"
+    echo "Script failed to download. Please run: 'sudo bash $SCRIPTS/nextcloud-startup-script.sh' again."
+    exit 1
 fi
 
 mv $SCRIPTS/index.php $WWW_ROOT/index.php && rm -f $WWW_ROOT/html/index.html
@@ -525,8 +495,8 @@ dpkg-reconfigure openssh-server
 
 # Generate new MySQL password
 echo
-bash "$SCRIPTS/change_mysql_pass.sh" && wait
-if [ $? -eq 0 ]; then
+bash "$SCRIPTS/change_mysql_pass.sh" && wait # is the exit status always 0 on if this is sucessfull?
+if [ $? -eq 0 ]; then # skip this?
   rm "$SCRIPTS/change_mysql_pass.sh"
   {
   echo "[mysqld]"
@@ -544,9 +514,9 @@ echo "Please be patient, it may take a while."
 sudo /etc/init.d/mysql restart
 RESULT="mysqlshow --user=root --password=$(cat $PW_FILE) $NCDB| grep -v Wildcard | grep -o $NCDB"
 if [ "$RESULT" == "$NCDB" ]; then
-    mysql -u root -e "ALTER DATABASE $NCDB CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;"
+    mysql -u root -e "ALTER DATABASE $NCDB CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;" # want to test if this is succesfull
 fi
-if [ $? -eq 0 ]
+if [ $? -eq 0 ] # Skip this?
 then
     sudo -u www-data $NCPATH/occ config:system:set mysql.utf8mb4 --type boolean --value="true"
     sudo -u www-data $NCPATH/occ maintenance:repair
@@ -700,10 +670,8 @@ service apache restart
 # Increase max filesize (expects that changes are made in /etc/php/7.0/apache2/php.ini)
 # Here is a guide: https://www.techandme.se/increase-max-file-size/
 VALUE="# php_value upload_max_filesize 513M"
-if grep -Fxq "$VALUE" $NCPATH/.htaccess
+if ! grep -Fxq "$VALUE" $NCPATH/.htaccess
 then
-        echo "Value correct"
-else
         sed -i 's/  php_value upload_max_filesize 513M/# php_value upload_max_filesize 513M/g' $NCPATH/.htaccess
         sed -i 's/  php_value post_max_size 513M/# php_value post_max_size 513M/g' $NCPATH/.htaccess
         sed -i 's/  php_value memory_limit 512M/# php_value memory_limit 512M/g' $NCPATH/.htaccess
